+++ spimdisasm/common/SymbolsSegment.py
@@ -352,8 +352,46 @@ class SymbolsSegment:
         0xA4940010: "USB0_STATUS_REG",
         0xA4A00000: "D_A4A00000",
         0xA4A40010: "USB1_STATUS_REG",
+
+        # libleo (64DD) address range
+        0xA5000508: "LEO_CMD",
+        0xA5000508: "LEO_STATUS",
+
+        0xA5000510: "LEO_BM_CTL",
+        0xA5000510: "LEO_BM_STATUS",
+
+        0xA5000518: "LEO_SEQ_CTL",
+        0xA5000518: "LEO_SEQ_STATUS",
+
+
+        0xA5000000: "LEO_C2_BUFF",      # C2 Sector Buffer
+        0xA5000400: "LEO_SECTOR_BUFF",  # Data Sector Buffer
+        0xA5000500: "LEO_DATA",         # Data
+        0xA5000504: "LEO_MISC_REG",     # Misc Register
+
+        0xA500050C: "LEO_CUR_TK",       # Current Track
+
+        0xA5000514: "LEO_ERR_SECTOR",   # Sector Error Status
+
+        0xA500051C: "LEO_CUR_SECTOR",   # Current Sector
+        0xA5000520: "LEO_HARD_RESET",   # Hard Reset
+        0xA5000524: "LEO_C1_S0",        # C1
+        0xA5000528: "LEO_HOST_SECBYTE", # Sector Size (in bytes)
+        0xA500052C: "LEO_C1_S2",        # C1
+        0xA5000530: "LEO_SEC_BYTE",     # Sectors per Block, Full Size
+        0xA5000534: "LEO_C1_S4",        # C1
+        0xA5000538: "LEO_C1_S6",        # C1
+        0xA500053C: "LEO_CUR_ADDR",     # Current Address?
+        0xA5000540: "LEO_ID_REG",       # ID
+        0xA5000544: "LEO_TEST_REG",     # Test Read
+        0xA5000548: "LEO_TEST_PIN_SEL", # Test Write
+        0xA5000580: "LEO_RAM_ADDR",     # Microsequencer RAM
     }
     "N64 OS hardware registers"
+
+    iQueHardwareReg: dict[int, str] = {
+    }
+    "iQue OS hardware registers"
 
 
     def fillLibultraSymbols(self) -> None: